+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Extract click methods to `ClickMethods` module
 * Increase test coverage of `ClickMethods` module
+* Remove `WebDriver#click_and_wait`
 
 ## 0.3.2 (2020-06-30)